backend/app/api/endpoints/chat.py
from fastapi import APIRouter
from pydantic import BaseModel
from app.services.rag_service import rag_service

router = APIRouter()

class ChatRequest(BaseModel):
    query: str

class ChatResponse(BaseModel):
    answer: str

@router.post("/chat", response_model=ChatResponse)
async def chat(request: ChatRequest):
    answer = await rag_service.chat(request.query)
    return ChatResponse(answer=answer)

