backend/requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-core
langchain
supabase==2.5.0
pypdf
python-multipart
python-dotenv
google-generativeai
